sureshBot/requirements.txt
PyNaCl==1.4.0
discord.py[voice]
youtube_dl==2021.12.17
async-timeout
dnspython==1.16.0
pytz
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib